Implemented the dynamic and clean way of importing all the icons/links and changing the color of the icons on the hover.
Using .map functionality I was first mapping through all the links inside api data, and then with ternary operator I checked if the name from mockedIcon array is equal to the name inside the api data, and only then show the icon, with this I'm able to show only the icons that are needed for certain artist.

Also with the combination of useState and onMouseOver function, I was able to change the color of the icon, by checking if the hover link is being hovered, and if it is change its color.