Made SubGenreVoting component that shows inside a pop up modal. Also implemented new way of changing buttons.
The reason I made a component for SubGenreVoting part is because it is being showed two times in the task, and so to escape the copying of the same code twice I made a component out of that. Also with this new way of changing types of button, we can now have multiple versions of buttons.